==> R markdown revision/08_citation_20240810.Rmd <==
---
title: "MAFLD"
output: html_document
date: "2024-06-17"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(bibliometrix)
library(tidyverse)
library(ggtext)
library(glue)
library(ggrepel)
library(countrycode)
library(sf)
library(rnaturalearth)
library(rnaturalearthdata)
library(rmapshaper)
library(CoordinateCleaner)
library(igraph)
library(ggraph)
library(tidygraph)
library(patchwork)
library(cowplot)
```

```{r}
# load workspace
load("rdata/07_reference.RData")
```

```{r}
# difference in CR_di between wos and scopus likely to be acceptable
wos_cr_map  %>% count(CR_di) %>%
  full_join(scopus_cr_map %>% count(CR_di), by = join_by(CR_di), suffix = c("_w", "_s")) %>%
  mutate(dif = abs(n_w - n_s)) %>%
  filter(dif >= 10) %>%
  arrange(desc(dif))
```

```{r}
# combine filtered references from wos and scopus
df_wos_cr <- wos_cr_map %>% select(!dplyr::ends_with("_w")) %>%
  bind_rows(scopus_cr_map %>% select(!dplyr::ends_with("_s"))) %>%
  distinct(UI, CR_di, .keep_all =  TRUE) %>%
  ungroup()
```

```{r}
# create edge list
# from = CR_di (cited ref), to = UI (citing paper)
df_cr_edgelist <- df_wos_cr %>%
  select(CR_di, UI) %>%
  count(CR_di, UI) %>%
  rename(from = "CR_di", to = "UI") %>%
  select(from, to, n) %>%
  filter(from != to)
```

```{r}
# check all edge weight is 1
df_cr_edgelist %>%
  count(from, to) %>%
  rename(weight = "n") %>%
  count(weight)
```

```{r}
# no self loop edge
df_cr_edgelist %>% filter(from == to)
```

```{r}
# construct direct citation network to detect multiple edges
gr_cr_dir <- graph_from_data_frame(df_cr_edgelist, directed = FALSE, vertices = NULL)

tbl_gr_cr_dir <- as_tbl_graph(gr_cr_dir, directed = FALSE)
```

```{r}
tbl_gr_cr_dir

# A tbl_graph: 1329 nodes and 7784 edges
# An undirected multigraph with 5 components
```

```{r}
# check self-loop edge
tbl_gr_cr_dir %>%
  activate(edges) %>%
  mutate(loop = edge_is_loop()) %>%
  as_tibble() %>%
  filter(loop == TRUE)
```

```{r}
# check multiple edges
tbl_gr_cr_dir %>%
  activate(edges) %>%
  mutate(multi = edge_is_multiple()) %>%
  as_tibble() %>%
  filter(multi == TRUE)
# "47 - 343"
```

```{r}
# check connections between 47 and 343
tbl_gr_cr_dir %>%
  activate(nodes) %>%
  mutate(row = row_number()) %>%
  filter(row %in% c(47, 343))

# name                        row
# <chr>                       <int>
# 10.1007/s12072-020-10094-2  47
# 10.1111/liv.14478           343
```

```{r}
# correct
df_wos_cr %>% filter(UI == "10.1007/s12072-020-10094-2" &
                 CR_di == "10.1111/liv.14478")

# wrong
df_wos_cr %>% filter(UI == "10.1111/liv.14478" &
                 CR_di == "10.1007/s12072-020-10094-2")
```

```{r}
# check wrong edge in df_cr_edgelist(from = "CR_di", to = "UI")
df_cr_edgelist %>%
  filter(from == "10.1007/s12072-020-10094-2" & to == "10.1111/liv.14478")

nrow(df_cr_edgelist)
#
```

```{r}
# remove wrong edge
df_cr_edgelist <- df_cr_edgelist %>%
  filter(!(from == "10.1007/s12072-020-10094-2" & to == "10.1111/liv.14478"))
nrow(df_cr_edgelist)
# 7778

# end of cleaning edgelist
```


```{r}
# visualise citation counts in web of science
df_bx %>%
  filter(!is.na(TC_w)) %>%
  slice_max(TC_w, n = 20) %>%
  mutate(PN2 = case_when(PM == "32278004" ~ paste(PN2, "[a]"),
                         .default = as.character(PN2))) %>%
  mutate(PN2 = case_when(PM == "32044314" ~ paste(PN2, "[b]"),
                         .default = as.character(PN2))) %>%
  mutate(PN2 = case_when(PM == "33006093" ~ paste(PN2, "[c]"),
                         .default = as.character(PN2))) %>%
  mutate(TI = str_remove(TI, "LETTER TO THE EDITOR: ")) %>%
  mutate(TI = str_to_sentence(TI)) %>%
  mutate(TI = str_replace(TI, "Mafld", "MAFLD")) %>%
  mutate(TI = str_replace(TI, "mafld", "MAFLD")) %>%
  mutate(TI = str_replace(TI, "nafld", "NAFLD")) %>%
  mutate(TI = str_replace(TI, "covid-19", "COVID-19")) %>%
  mutate(TI = str_replace(TI, "ckd", "CKD")) %>%
  mutate(TI = str_replace(TI, "asian pacific association for the study of the liver", "Asian Pacific Association for the Study of the Liver")) %>%
  ggplot(aes(TC_w, fct_reorder(PN2, TC_w))) +
  geom_col(fill = "#4daf4a", alpha = 0.5)+
  geom_text(aes(x = 40, label = TI), hjust = 0, vjust = 0.5, size = 3)+
  scale_x_continuous(breaks = seq(0, 2000, 500), 
                   limits = c(0,2000), 
                   expand = c(0,0))+
  labs(x = "Citation count", y = "Publication", title = "Web of Science") +
  theme_classic()+
  theme(plot.margin = unit(c(0,0.5,0,0), "cm"))
```

```{r}
# visualise citation counts in scopus
df_bx %>%
  filter(!is.na(TC_s)) %>%
  slice_max(TC_s, n = 20) %>%
  mutate(PN2 = case_when(PM == "32278004" ~ paste(PN2, "[a]"),
                         .default = as.character(PN2))) %>%
  mutate(PN2 = case_when(PM == "32044314" ~ paste(PN2, "[b]"),
                         .default = as.character(PN2))) %>%
  mutate(PN2 = case_when(PM == "33006093" ~ paste(PN2, "[c]"),
                         .default = as.character(PN2))) %>%
  mutate(TI = str_remove(TI, "LETTER TO THE EDITOR: ")) %>% 
  mutate(TI = str_to_sentence(TI)) %>%
  mutate(TI = str_remove(TI, "LETTER TO THE EDITOR: ")) %>%
  mutate(TI = str_to_sentence(TI)) %>%
  mutate(TI = str_replace(TI, "Mafld", "MAFLD")) %>%
  mutate(TI = str_replace(TI, "mafld", "MAFLD")) %>%
  mutate(TI = str_replace(TI, "nafld", "NAFLD")) %>%
  mutate(TI = str_replace(TI, "covid-19", "COVID-19")) %>%
  mutate(TI = str_replace(TI, "ckd", "CKD")) %>%
  mutate(TI = str_replace(TI, "asian pacific association for the study of the liver", "Asian Pacific Association for the Study of the Liver")) %>%
  ggplot(aes(TC_s, fct_reorder(PN2, TC_s))) +
  geom_col(fill = "#F9A134", alpha = 0.5)+
  geom_text(aes(x = 40, label = TI), hjust = 0, vjust= 0.5, size = 3)+
  scale_x_continuous(breaks = seq(0, 2000, 500), 
                   limits = c(0,2000), 
                   expand = c(0,0))+
  labs(x = "Citation count", y = "Publication", title = "Scopus") +
  theme_classic()+
  theme(plot.margin = unit(c(0,0.5,0,0), "cm"))
```

```{r}
# visualise citation counts from MAFLD literature
df_cr_edgelist %>% count(from) %>%
  left_join(df_bx %>% select(TI, PY, PM, PN2, PN3, UI), by = join_by(from == UI)) %>%
  slice_max(n, n = 20) %>%
  mutate(PN2 = case_when(PM == "32278004" ~ "Eslam M (2020) [a]",
                   PM == "32044314" ~ "Eslam M (2020) [b]",
                   PM == "33006093" ~ "Eslam M (2020) [c]",
                   PM == "33031758" ~ "Shiha G (2021) [a]",
                   PM == "33181119" ~ "Shiha G (2021) [b]",
                   .default = as.character(PN2))) %>%
  mutate(TI = str_to_sentence(TI)) %>%
  mutate(TI = str_replace(TI, "Mafld", "MAFLD")) %>%
  mutate(TI = str_replace(TI, "mafld", "MAFLD")) %>%
  mutate(TI = str_replace(TI, "nafld", "NAFLD")) %>%
  mutate(TI = str_replace(TI, "covid-19", "COVID-19")) %>%
  mutate(TI = str_replace(TI, "ckd", "CKD")) %>%
  mutate(TI = str_replace(TI, "asian pacific association for the study of the liver", "Asian Pacific Association for the Study of the Liver")) %>%
  ggplot(aes(n, fct_reorder(PN2, n)))+
  geom_col(fill = "#377eb8", alpha = 0.5)+
  geom_text(aes(x = 20, label = TI), hjust = 0, size = 3)+
  scale_x_continuous(breaks = seq(0, 1000, 200), 
                   limits = c(0,1000), 
                   expand = c(0,0))+
  labs(x = "Citation count", y = "Publication") +
  theme_classic()+
  theme(plot.margin = unit(c(0,5,0,0), "mm"))
```

```{r}
# use scopus for plotting total ciation counts (Fig 4a)
p4_a <- df_bx %>%
  filter(!is.na(TC_s)) %>%
  slice_max(TC_s, n = 15) %>%
  mutate(PN2 = case_when(PM == "32278004" ~ paste(PN2, "[a]"),
                         .default = as.character(PN2))) %>%
  mutate(PN2 = case_when(PM == "32044314" ~ paste(PN2, "[b]"),
                         .default = as.character(PN2))) %>%
  mutate(PN2 = case_when(PM == "33006093" ~ paste(PN2, "[c]"),
                         .default = as.character(PN2))) %>%
  mutate(TI = str_remove(TI, "LETTER TO THE EDITOR: ")) %>% 
  mutate(TI = str_to_sentence(TI)) %>%
  mutate(TI = str_remove(TI, "LETTER TO THE EDITOR: ")) %>%
  mutate(TI = str_to_sentence(TI)) %>%
  mutate(TI = str_replace(TI, "Mafld", "MAFLD")) %>%
  mutate(TI = str_replace(TI, "mafld", "MAFLD")) %>%
  mutate(TI = str_replace(TI, "nafld", "NAFLD")) %>%
  mutate(TI = str_replace(TI, "covid-19", "COVID-19")) %>%
  mutate(TI = str_replace(TI, "ckd", "CKD")) %>%
  mutate(TI = str_replace(TI, "asian pacific association for the study of the liver", "Asian Pacific Association for the Study of the Liver")) %>%
  mutate(TI = str_replace(TI, "kupffer", "Kupffer")) %>%
  ggplot(aes(TC_s, fct_reorder(PN2, TC_s))) +
  geom_col(fill = "#ff7f00", alpha = 0.5, width = 0.7)+
  geom_text(aes(x = 2000/50, label = TI), hjust = 0, vjust= 0.5, size = 2.8)+
  scale_x_continuous(breaks = seq(0, 2000, 500), 
                   limits = c(0,2000), 
                   expand = c(0,0))+
  labs(x = "Citation count in Scopus", y = "Publication", tag = "a") +
  theme_classic()+
  theme(plot.tag = element_text(size = 12, face="bold"),
         axis.title = element_text(size = 8, face = "bold"),
         axis.text.x = element_text(size = 8, colour = "black"),
         axis.text.y = element_text(size = 8, colour = "black"),
         axis.line = element_line(linewidth = 0.3, colour = "black"),
         axis.ticks = element_line(linewidth = 0.3, colour = "black"),
         plot.margin = unit(c(0,5,0,0), "mm"))
```

```{r}
# plot citation counts from MAFLD literature (Fig 4b)
p4_b <- df_cr_edgelist %>% 
  count(from) %>%
  left_join(df_bx %>% select(TI, PY, PM, PN2, PN3, UI), by = join_by(from == UI)) %>%
  slice_max(n, n = 15) %>%
  mutate(PN2 = case_when(PM == "32278004" ~ "Eslam M (2020) [a]",
                   PM == "32044314" ~ "Eslam M (2020) [b]",
                   PM == "33006093" ~ "Eslam M (2020) [c]",
                   .default = as.character(PN2))) %>%
  mutate(TI = str_to_sentence(TI)) %>%
  mutate(TI = str_replace(TI, "Mafld", "MAFLD")) %>%
  mutate(TI = str_replace(TI, "mafld", "MAFLD")) %>%
  mutate(TI = str_replace(TI, "nafld", "NAFLD")) %>%
  mutate(TI = str_replace(TI, "covid-19", "COVID-19")) %>%
  mutate(TI = str_replace(TI, "ckd", "CKD")) %>%
  mutate(TI = str_replace(TI, "asian pacific association for the study of the liver", "Asian Pacific Association for the Study of the Liver")) %>%
  mutate(TI = str_replace(TI, "latin american association for the study of the liver", "Latin American Association for the Study of the Liver")) %>%
  mutate(TI = str_replace(TI, "aleh", "ALEH")) %>%
  ggplot(aes(n, fct_reorder(PN2, n)))+
  geom_col(fill = "#377eb8", alpha = 0.5, width = 0.7)+
  geom_text(aes(x = 2000/50, label = TI), hjust = 0, size = 2.8)+
  scale_x_continuous(breaks = seq(0, 2000, 500), 
                   limits = c(0,2000), 
                   expand = c(0,0))+
  labs(x = "Citation count from MAFLD literature", y = "Publication", tag = "b") +
  theme_classic()+
   theme(plot.tag = element_text(size = 12, face="bold"),
         axis.title = element_text(size = 8, face = "bold"),
         axis.text.x = element_text(size = 8, colour = "black"),
         axis.text.y = element_text(size = 8, colour = "black"),
         axis.line = element_line(linewidth = 0.3, colour = "black"),
         axis.ticks = element_line(linewidth = 0.3, colour = "black"),
         plot.margin = unit(c(0,5,0,0), "mm"))
```

```{r}
# combine plot and save
#p4 <- p4_a + p4_b + plot_layout(ncol = 1, nrow = 2)
#ggsave("figure/Fig3.pdf", p4, width = 174, height = 150, unit = "mm")
```

```{r}
# citation trend
df_cr_edgelist <- df_cr_edgelist %>%
  rename(CR = "from", UI = "to")
# all 1
```

```{r}
colnames(df_de_long)
```

```{r}
# reference trend

df_cr_trend <- df_cr_edgelist %>%
  left_join(df_bx %>% select(UI, PM, DI),
            by = join_by(CR == UI))

```

```{r}
# add cited year

df_cr_trend <- df_cr_trend %>%
  left_join(df_bx %>% select(UI, PM, PY), suffix = c("_cr", "_ui"),
            by = join_by(UI))
```

```{r}
df_cr_trend %>% filter(PM_cr == "32278004") %>% 
    mutate(PY = case_when(PY == "2024" ~ "2023",
                   .default = as.character(PY))) %>%
  count(PY)
```


```{r}
df_cr_trend %>% filter(PM_cr == "32044314") %>%
  mutate(PY = case_when(PY == "2024" ~ "2023",
                   .default = as.character(PY))) %>%
  count(PY) %>%
  ggplot(aes(PY, n, group = 1))+
  geom_line()
```

```{r}
# jhep 32278004 Eslam M (2020) [a]
# ge 32044314 Eslam M (2020) [b]
# apasl 33006093 Eslam M (2020) [c]
# comparison 32478487
# fibrosis 32997882
# CV 33348045 (34129272)
# all cause 34380057
# COVID (32414813) 32320741 
# CKD 33212070

cr_trend_pm <-  c(32278004, 32044314, 33006093, 32478487, 32997882, 33348045, 34380057, 32320741, 33212070)
```

```{r}

df_cr_trend %>%
  mutate(PY = case_when(PY == "2024" ~ "2023",
                        .default = as.character(PY))) %>%
  filter(PM_cr %in% cr_trend_pm) %>%
  count(PY, CR) 
  
```

```{r}
cr_trend_order_pm <-  df_cr_trend %>%
  mutate(PY = case_when(PY == "2024" ~ "2023",
                        .default = as.character(PY))) %>%
  filter(PM_cr %in% cr_trend_pm) %>%
  count(PY, PM_cr) %>%
  filter(PY == "2023") %>%
  arrange(desc(n)) %>%
  pull(PM_cr)

length(cr_trend_order_pm)
cr_trend_order_pm
```

```{r}
df_cr_trend_plot <- df_cr_trend %>%
  mutate(PY = case_when(PY == "2024" ~ "2023",
                        .default = as.character(PY))) %>%
  filter(PM_cr %in% cr_trend_pm) %>%
  count(PY, CR) %>%
  left_join(df_bx %>% select(PM, PN2, UI), by = join_by(CR == UI)) %>%
  mutate(PN2 = case_when(PM == "32278004" ~ "Eslam M (2020) [a]",
                   PM == "32044314" ~ "Eslam M (2020) [b]",
                   PM == "33006093" ~ "Eslam M (2020) [c]",
                   .default = as.character(PN2)))
```

```{r}
cr_trend_order_pn2 <- df_cr_trend_plot %>% filter(PY == "2023") %>% arrange(desc(n)) %>% pull(PN2)
cr_trend_order_pn2
```

```{r}
#df_cr_trend_plot <- 
p4_c <-  df_cr_trend_plot %>%
  mutate(PN2 = factor(PN2, levels = cr_trend_order_pn2)) %>%
  ggplot(aes(PY, n, group = PN2, colour = PN2))+
  geom_line(linewidth = 0.3)+
   geom_point(size = 1)+
    scale_color_manual(values = c("#377eb8",
                                  "#ff7f00",
                                  "#e41a1c",
                                  "#4daf4a",
                                  "#a65628",
                                  "#f781bf",
                                  "#999999",
                                  "black",
                                  "#984ea3"),
                       name = "Reference")+
  guides(colour = guide_legend(ncol = 2))+
  scale_y_continuous(breaks = c(0, 10, 50, 100, 200, 300),
                     limit= c(0, 300),
                     trans = "log1p",
                     expand = c(0, 0))+
#  annotation_logticks(sides = "l", outside = TRUE) +
  labs(x = "Year", y = "Cited count", tag = "c") +
  coord_cartesian(clip = "off")+
  theme_classic()+
   theme(plot.tag = element_text(size = 12, face="bold"),
         axis.title = element_text(size = 8, colour = "black", face = "bold"),
         axis.text.x = element_text(size = 8, colour = "black"),
         axis.text.y = element_text(size = 8, colour = "black"),
         axis.line = element_line(linewidth = 0.3, colour = "black"),
         axis.ticks = element_line(linewidth = 0.3, colour = "black"),
         legend.title = element_text(size = 8, colour = "black", face = "bold"),
         legend.text = element_text(size = 8, colour = "black"), 
         plot.margin = unit(c(0,0,0,0), "mm"))

p4_c
```



```{r}
df_de_long %>% filter(is.na(DE)) %>% sum()
```

```{r}
df_de_long %>% filter(UI == "10.2147/dddt.s417378") %>% pull(DE)
```


```{r}
rows_to_list <- function(x){
  de <- df_de_long %>% filter(UI == x) %>% pull(DE) %>% list()
  df <- tibble(UI = x, DE = de)
  return(df)
} 
```

```{r}
ui_df_de <- df_de %>% pull(UI) 

ui_df_de %>% head()
```

```{r}
rows_to_list("10.2147/dddt.s417378")
```


```{r}
df_de_list <- map(ui_df_de, rows_to_list) |> list_rbind()
```

```{r}
df_de_list
```


```{r}
# citation ternd

df_de_list <- df_de_list %>%
  left_join(df_de %>% select(-c(DE, PY)),
            by = join_by(UI))
```


```{r}
# join nested df_de to df_cr_edgelist
df_de_trend <- df_cr_edgelist %>% 
  #rename(UI = "to") %>%
  #rename(CR = "from") %>%
  left_join(df_de_list, by = join_by(CR == UI))
```

```{r}
# add cited year
# left join PY of UI (published year of citing paper)
df_de_trend <- df_de_trend %>%
  left_join(df_de %>% select(UI, PY), by = join_by(UI) )
```


```{r}
df_de_trend_long <- df_de_trend %>% unnest_longer(DE) 
```

```{r}
colnames(df_de_trend_long)
```

```{r}
# clean keywords for plotting
df_de_trend_long <- df_de_trend_long %>%
  group_by(UI) %>%
  distinct(DE, .keep_all = TRUE) %>%
  ungroup() %>%
  filter(DE %in% c("MAFLD",
                   "OBESITY",
                   "METABOLIC SYNDROME",
                   "TYPE 2 DIABETES MELLITUS",
                   "LIVER FIBROSIS",
                   "HEPATOCELLULAR CARCINOMA",
                   "CARDIOVASCULAR DISEASE",
                   "COVID-19",
                   "GUT MICROBIOTA")) %>%
  mutate(DE = str_to_sentence(DE)) %>%
  mutate(DE = case_when(DE == "Mafld" ~ "MAFLD",
                        DE == "Type 2 diabetes mellitus" ~ "Type 2 diabetes",
                        DE == "Hepatocellular carcinoma" ~ "HCC",
                        DE == "Covid-19" ~ "COVID-19",
                        .default = as.character(DE)))  %>%
  mutate(PY = case_when(PY == "2024" ~ "2023",
                          .default = as.character(PY)))
```

```{r}
de_trend_order <- df_de_trend_long %>% count(PY, DE) %>% filter(PY == "2023") %>% arrange(desc(n)) %>% pull(DE)

de_trend_order
```


```{r}
p4_d <- df_de_trend_long %>%
  count(DE, PY) %>%
  mutate(DE = factor(DE, levels = de_trend_order )) %>%
  ggplot(aes(PY, n, group = DE, colour = DE )) +
    geom_line(linewidth = 0.3)+
    geom_point(size = 1)+
    scale_color_manual(values = c("#377eb8",
                                  "#ff7f00",
                                  "#e41a1c",
                                  "#4daf4a",
                                  "#a65628",
                                  "#f781bf",
                                  "#999999",
                                  "black",
                                  "#984ea3"),
                       name = "Keyword")+
   guides(colour = guide_legend(ncol = 2))+
  scale_y_continuous(breaks = c(0,10, 50, 100, 250, 500),
                     limit= c(0, 500),
                     trans = "log1p",
                     expand = c(0, 0))+
  labs(x = "Year", y = "Cited count", tag = "d") +
  theme_classic()+
   theme(plot.tag = element_text(size = 12, face="bold"),
         axis.title = element_text(size = 8, colour = "black", face = "bold"),
         axis.text.x = element_text(size = 8, colour = "black"),
         axis.text.y = element_text(size = 8, colour = "black"),
         axis.line = element_line(linewidth = 0.3, colour = "black"),
         axis.ticks = element_line(linewidth = 0.3, colour = "black"),
         legend.title = element_text(size = 8, colour = "black", face = "bold"),
         legend.text = element_text(size = 8, colour = "black"), 
         plot.margin = unit(c(0,0,0,0), "mm"))
  
```


```{r}
p4_d
```


```{r}
p4 <- p4_a + p4_b + free(p4_c) + free(p4_d) + plot_layout(nrow = 4, ncol = 1,
                                                          heights = c(4, 4, 3, 3)  )

ggsave("figure/Fig3.pdf", plot = p4, width = 174, height = 234, unit = "mm")
```

```{r}
save.image("rdata/08_citation.Rdata")
```
